moved some styles to seperate files, did some work on readme
