+++ README.md
@@ -1,4 +1,91 @@
-# Know Thyself
 
-Start back end in root folder with ```npm run dev```
-Start front end in frontend folder with ```npm start```
+<p align="center">
+  <a href="https://gitpoint.co/">
+    <img alt="Know Thyself" title="Know Thyself" src="https://i.imgur.com/Q3qbgNu.png" width="450">
+  </a>
+</p>
+
+<p align="center">
+  Simple Journaling - Hard Prompts
+</p>
+
+
+Know Thyself if a web journaling app to help build self knowledge through the practice of responding to thought provoking writing prompts with a minimum word count.
+
+## Features
+  - Login with Google or make a new account locally
+  - See the weather and time in your area
+  - Get supplied with thought provoking prompts for journaling
+  - Ability to write words on said prompt
+  - Complete at least 500 words on that prompt to save it
+  - View entries and prompts from past days
+
+## Getting Started / Installing
+#### [Use our App on Heroku](https://dry-cove-74246.herokuapp.com)
+----------or-------------
+#### Install locally following these steps
+- Clone or download the repo
+- In the root folder... ```npm install```.
+- In knowthyself/frontend... ```npm install```
+
+This will install all the dependencies needed to run locally.
+
+To run locally you need to...
+- Have two instances of terminal open
+- In one, in the root folder...```npm run dev```
+- In the second, in knowthyself/frontend... ```npm start```
+
+
+
+## Running the tests
+
+``mocha tesing/test.js``
+
+### Break down into end to end tests
+
+Explain what these tests test and why
+
+```
+Give an example
+```
+
+### And coding style tests
+
+Explain what these tests test and why
+
+```
+Give an example
+```
+
+
+
+## Built With
+
+* [Dropwizard](http://www.dropwizard.io/1.0.2/docs/) - The web framework used
+* [Maven](https://maven.apache.org/) - Dependency Management
+* [ROME](https://rometools.github.io/rome/) - Used to generate RSS Feeds
+
+## Contributing
+
+Please read [CONTRIBUTING.md](https://gist.github.com/PurpleBooth/b24679402957c63ec426) for details on our code of conduct, and the process for submitting pull requests to us.
+
+## Versioning
+
+We use [SemVer](http://semver.org/) for versioning. For the versions available, see the [tags on this repository](https://github.com/your/project/tags).
+
+## Authors
+
+* **Jake** - *Initial work* - [PurpleBooth](https://github.com/PurpleBooth)
+* **Alison** - *Initial work* -
+* ** Kyle** - *Initial work* -
+
+
+## License
+
+This project is licensed under the MIT License - see the [LICENSE.md](LICENSE.md) file for details
+
+## Acknowledgments
+
+* Hat tip to anyone whose code was used
+* Inspiration
+* Ed
