--- frontend/src/navbar/Navbar.css
@@ -1,3 +1,0 @@
-.toolbar {
-  background-color: #373737;
-}
