+++ package.json
@@ -39,6 +39,7 @@
     "passport-google-token": "^0.1.2",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
+    "puppeteer": "^1.10.0",
     "request": "^2.88.0",
     "validator": "^10.8.0"
   }
